# backend/openapi.json
{
  "name": "n-verify-api",
  "version": "1.0.0",
  "description": "N-Verify Protocol API Documentation",
  "openapi": "3.0.0",
  "info": {
    "title": "N-Verify Protocol API",
    "version": "1.0.0",
    "description": "API for verifying AI outputs with cryptographic proofs"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "organizationName": { "type": "string" },
                  "role": { "type": "string", "enum": ["user", "validator"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
    "/verify": {
      "get": {
        "summary": "List verification requests",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of verifications"
          }
        }
      },
      "post": {
        "summary": "Create verification request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "domain": { "type": "string", "enum": ["medical", "legal", "financial"] },
                  "aiOutput": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification created"
          }
        }
      }
    },
    "/verify/{requestId}": {
      "get": {
        "summary": "Get verification status",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification details"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}